Migrate resource validation and associated methods out of authenticated middleware
