+++ shell/components/ResourceDetail/index.vue
@@ -14,6 +14,7 @@ import { clone, diff } from '@shell/utils/object';
 import IconMessage from '@shell/components/IconMessage';
 import ForceDirectedTreeChart from '@shell/components/fleet/ForceDirectedTreeChart';
 import { checkSchemasForFindAllHash } from '@shell/utils/auth';
+import ResourceValidation from '@shell/mixins/resource-validation';
 
 function modeFor(route) {
   if ( route.query?.mode === _IMPORT ) {
@@ -48,7 +49,7 @@ export default {
     IconMessage,
   },
 
-  mixins: [CreateEditView],
+  mixins: [ResourceValidation, CreateEditView],
 
   props: {
     storeOverride: {
